Strip the hstspreload server to its basics.
The Google Cloud test project is hstspreload-mvm.
Managed VMs allow us to eventually host this code at hstspreload.appspot.com but make tls.Dial calls from Go.
