+++ Makefile
@@ -1,0 +1,7 @@
+.PHONY: serve
+serve:
+	go run server.go
+
+.PHONY: deploy
+deploy:
+	DYLD_INSERT_LIBRARIES="" aedeploy gcloud preview app deploy app.yaml --promote
