--- msg.html
@@ -1,70 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-	<head>
-		<title>HSTS Message</title>
-		<style>
-body {
-	margin-top: 2cm;
-}
-#entry {
-	padding: 2cm;
-	margin: auto;
-	border: 5px solid gray;
-	background: #dfe;
-	border-radius: 30px;
-	width: 50em;
-}
-
-#entry h2 {
-	font-family: sans-serif;
-	text-align: center;
-}
-
-#entry input {
-	font-family: sans-serif;
-	margin: auto;
-	font-size: 1.5em;
-	display: block;
-	margin-bottom: 1em;
-}
-
-#entry p {
-	font-family: sans-serif;
-	font-size: 1.2em;
-	display: block;
-	text-align: center;
-}
-
-#error {
-	color: red;
-}
-
-#text {
-	padding: 2cm;
-	margin: auto;
-	width: 45em;
-	font-size: 1.2em;
-	color: #333;
-	font-family: sans-serif;
-	line-height: 140%;
-	text-align: justify;
-}
-
-#text p {
-	padding-bottom: 1em;
-}
-		</style>
-
-	</head>
-
-	<body>
-		<div id="entry">
-			<form method="POST" action="/setmessage">
-				<input type="text" placeholder="example.com" name="domain" autofocus></input>
-				<input type="text" name="msg"></input>
-				<input type="submit" name="submit"></input>
-			</form>
-		</div>
-	</body>
-</html>
